+++ src/main/java/dev/zenith/offlineuuid/OfflineUUIDConfig.java
@@ -5,6 +5,21 @@ package dev.zenith.offlineuuid;
  */
 public class OfflineUUIDConfig {
     public boolean enabled = false;
-    public String prefix = "OfflinePlayer";
-    public String offlineUuid = null;
+    public SideConfig server = new SideConfig();
+    public SideConfig client = new SideConfig();
+
+    public static final class SideConfig {
+        public boolean enabled = false;
+        public Mode mode = Mode.GENERATED;
+        public String fixedUuid = null;
+        public boolean addPrefix = true;
+        public String prefix = "OfflinePlayer:";
+    }
+
+    public enum Mode {
+        ORIGINAL,
+        FIXED,
+        RANDOM,
+        GENERATED
+    }
 }